datagen/pyproject.toml
[project]
name = "fractal-datagen"
version = "0.1.0"
description = "Fractal flame training data generation pipeline"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pillow>=10.0",
    "scipy>=1.11",
    "numba>=0.59",
    "tqdm>=4.66",
    "click>=8.1",
    "imagehash>=4.3",
]

[tool.setuptools.packages.find]
include = ["flamegen*", "preprocessing*"]

[project.scripts]
flamegen = "flamegen.generate_batch:main"
preprocess = "preprocessing.build_dataset:main"
